Moved feature_indices to a separate file
